requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
pgvector==0.2.4
sqlalchemy==2.0.23
alembic==1.13.0

# Vector embeddings
openai==1.3.7
cohere==4.37
voyageai==0.1.8
sentence-transformers==2.2.2

# Document processing
PyMuPDF==1.23.8
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2
pandas==2.1.4
numpy==1.24.3

# Email processing
email-validator==2.1.0
python-dotenv==1.0.0

# Async task queue
celery==5.3.4
redis==5.0.1

# Configuration and utilities
pyyaml==6.0.1
python-dotenv==1.0.0
click==8.1.7

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# File handling
aiofiles==23.2.1
pathlib2==2.3.7

# Text processing
nltk==3.8.1
spacy==3.7.2
textstat==0.7.3

# Security
cryptography==41.0.8
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Optional: Local LLM support
transformers==4.36.0
torch==2.1.1
accelerate==0.25.0 